Generalized DownloadableDocumentListSection and implemented download documents in AgreementRead
